fix: 라우팅 살짝 수정 + query 접근 코드 추가
