+++ project01_page/src/pages/one-bite/book/[id].tsx
@@ -1,0 +1,11 @@
+import { useRouter } from 'next/router'
+
+const BookDetail = () => {
+  const {
+    query: { id },
+  } = useRouter()
+  console.log(id)
+  return <div>Book Detail</div>
+}
+
+export default BookDetail
