--- project01_page/src/pages/one-bite/books/[id]/index.tsx
@@ -1,5 +1,0 @@
-const BookDetail = () => {
-  return <div>Book Detail</div>
-}
-
-export default BookDetail
