--- project01_page/src/pages/one-bite/search.tsx
@@ -1,4 +1,0 @@
-const Search = () => {
-  return <div>Search Page</div>
-}
-export default Search
