+++ project01_page/src/pages/one-bite/search/index.tsx
@@ -1,0 +1,9 @@
+import { useRouter } from 'next/router'
+
+const Search = () => {
+  const { query } = useRouter()
+  console.log(query)
+
+  return <div>Search Page</div>
+}
+export default Search
